Add directories for pruners.
